requirements.txt
django==1.9.10
Pillow==4.2.1
requests==2.11.1
mysqlclient
gunicorn==19.6.0
gevent==1.1.2
django-material==0.9.1
django-simple-history==1.8.1
django-extensions==1.7.4
django-allauth==0.30.0
django-smuggler==0.7.0
django-model-utils==3.1.2
django-debug-toolbar==1.5
django-cors-headers==1.2.0
easy-thumbnails==2.3
django-rest-swagger==2.1.1
djangorestframework==3.5.3
Markdown==2.6.7
django-filter==1.1.0
django-oauth-toolkit==0.10.0
django-rest-auth==0.9.2
exponent-server-sdk==0.1.1
numpy==1.13.3
pandas==0.21.0
django-pandas==0.4.5
django-filer==1.3.0